==> 4.10.questions.ipynb <==
# import pandas
import pandas as pd

# load cancer data into data frame
df = pd.read_csv('cancer_data.csv')
# display first five rows of data
df.head()

# print the column labels in the dataframe
for i, v in enumerate(df.columns):
    print(i, v)
