==> 4.15.assessing-quiz-solutions.ipynb <==
# Assessing and Building Intuition Quiz
Use the space below to explore `census_income_data.csv` to answer the quiz questions below.

# Import pandas
import pandas as pd

# Load census income data
df = pd.read_csv('census_income_data.csv')

# Work to answer the quiz questions
df.shape

df.info()

df.education.nunique()

df.describe()
