==> 5.05.assessing.ipynb <==
# Assessing
Use the space below to explore `winequality-red.csv` and `winequality-white.csv` to answer the quiz questions below.

import pandas as pd

red_df = pd.read_csv('winequality-red.csv', sep=';')
white_df = pd.read_csv('winequality-white.csv', sep=';')

print(red_df.shape)
red_df.head()

print(white_df.shape)
white_df.head()

red_df.isnull().sum()

white_df.isnull().sum()

white_df.duplicated().sum()

red_df.quality.nunique()

white_df.quality.nunique()

red_df.density.mean()
