==> 5.07.appending-solutions.ipynb <==
# Appending Data
First, import the necessary packages and load `winequality-red.csv` and `winequality-white.csv`.

# import numpy and pandas
import numpy as np
import pandas as pd

# load red and white wine datasets
red_df = pd.read_csv('winequality-red.csv', sep=';')
white_df = pd.read_csv('winequality-white.csv', sep=';')

red_df.rename(columns={'total_sulfur-dioxide':'total_sulfur_dioxide'}, inplace=True)

## Create Color Columns
Create two arrays as long as the number of rows in the red and white dataframes that repeat the value â€œredâ€ or â€œwhite.â€ NumPy offers really easy way to do this. Hereâ€™s the documentation for [NumPyâ€™s repeat](https://docs.scipy.org/doc/numpy/reference/generated/numpy.repeat.html) function. Take a look and try it yourself.

# create color array for red dataframe
color_red = np.repeat('red', red_df.shape[0])

# create color array for white dataframe
color_white = np.repeat('white', white_df.shape[0])

Add arrays to the red and white dataframes. Do this by setting a new column called 'color' to the appropriate array. The cell below does this for the red dataframe.

red_df['color'] = color_red
red_df.head()

Do the same for the white dataframe and use `head()` to confirm the change.

white_df['color'] = color_white
white_df.head()

## Combine DataFrames with Append
Check the documentation for [Pandas' append](https://pandas.pydata.org/pandas-docs/stable/generated/pandas.DataFrame.append.html) function and see if you can use this to figure out how to combine the dataframes. (Bonus: Why aren't we using the [merge](https://pandas.pydata.org/pandas-docs/stable/generated/pandas.DataFrame.merge.html) method to combine the dataframes?) If you donâ€™t get it, Iâ€™ll show you how afterwards. Make sure to save your work in this notebook! You'll come back to this later.

# append dataframes
wine_df = red_df.append(white_df) 

# view dataframe to check for success
wine_df.head()

## Save Combined Dataset
Save your newly combined dataframe as `winequality_edited.csv`. Remember, set `index=False` to avoid saving with an unnamed column!

wine_df.to_csv('winequality_edited.csv', index=False)
