==> 5.13.conclusions_groupby.ipynb <==
# Drawing Conclusions Using Groupby

# Load `winequality_edited.csv`
import pandas as pd

df = pd.read_csv('winequality_edited.csv')

### Is a certain type of wine associated with higher quality?

# Find the mean quality of each wine type (red and white) with groupby
df.groupby('color').mean().quality

### What level of acidity receives the highest average rating?

# View the min, 25%, 50%, 75%, max pH values with Pandas describe
df.describe().pH

# Bin edges that will be used to "cut" the data into groups
bin_edges = [2.72, 3.11, 3.21, 3.32, 4.01] # Fill in this list with five values you just found

# Labels for the four acidity level groups
bin_names = ['high', 'mod_high', 'medium', 'low'] # Name each acidity level category

# Creates acidity_levels column
df['acidity_levels'] = pd.cut(df['pH'], bin_edges, labels=bin_names)

# Checks for successful creation of this column
df.head()

# Find the mean quality of each acidity level with groupby
df.groupby('acidity_levels').mean().quality

# Save changes for the next section
df.to_csv('winequality_edited.csv', index=False)
