==> 5.19.wine_visualizations.ipynb <==
# Plotting with Matplotlib
Use Matplotlib to create bar charts that visualize the conclusions you made with groupby and query.

# Import necessary packages and load `winequality_edited.csv`


df =

### #1: Do wines with higher alcoholic content receive better ratings?
Create a bar chart with one bar for low alcohol and one bar for high alcohol wine samples. This first one is filled out for you.

# Use query to select each group and get its mean quality
median = df['alcohol'].median()
low = df.query('alcohol < {}'.format(median))
high = df.query('alcohol >= {}'.format(median))

mean_quality_low = low['quality'].mean()
mean_quality_high = high['quality'].mean()

# Create a bar chart with proper labels
locations = [1, 2]
heights = [mean_quality_low, mean_quality_high]
labels = ['Low', 'High']
plt.bar(locations, heights, tick_label=labels)
plt.title('Average Quality Ratings by Alcohol Content')
plt.xlabel('Alcohol Content')
plt.ylabel('Average Quality Rating');

### #2: Do sweeter wines receive higher ratings?
Create a bar chart with one bar for low residual sugar and one bar for high residual sugar wine samples.

# Use query to select each group and get its mean quality



# Create a bar chart with proper labels



### #3: What level of acidity receives the highest average rating?
Create a bar chart with a bar for each of the four acidity levels.

# Use groupby to get the mean quality for each acidity level


# Create a bar chart with proper labels



### Bonus: Create a line plot for the data in #3
You can use pyplot's [plot](https://matplotlib.org/stable/api/_as_gen/matplotlib.pyplot.plot.html) function for this.

Compare this with the bar chart. How might showing this visual instead of the bar chart affect someone's conclusion about this data?
