==> 6.05.Assess Data.ipynb <==

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
%matplotlib inline


df08 = pd.read_csv("all_alpha_08.csv")

df18 = pd.read_csv("all_alpha_18.csv")

#Check df dimensions
df08.shape


df18.shape


#Check head
df08.head(10)


df18.head(10)

#Check column types
df08.dtypes


type(df08['Cyl'][0])

type(df08['City MPG'][0])

type(df08['Greenhouse Gas Score'][0])

type(df18['Cyl'][0])

type(df18['City MPG'][0])

type(df18['Greenhouse Gas Score'][0])

df08.Fuel.unique()

df18.Fuel.unique()

#Check unique values
df08.nunique()


df18.nunique()


#Check for duplis
sum(df08.duplicated())


sum(df18.duplicated())


df08.info()


df18.info()


df08.isna().sum()

df18.isna().sum()
