==> 6.08 Inspecting Data types.ipynb <==
# load datasets
import pandas as pd

df_08 = pd.read_csv('data_08_v2.csv')
df_18 = pd.read_csv('data_18_v2.csv')


df_08.head(10)

df_18.head(2)

df_08.dtypes

df_18.dtypes

type(df_08['air_pollution_score'][0])

type(df_08['greenhouse_gas_score'][0])

type(df_18['greenhouse_gas_score'][0])

type(df_18['cmb_mpg'][0])
